Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -3,7 +3,7 @@
 int main() {
     float r;
     scanf("%f",&r);
-    int area = 3.14 * r * r;
+    float area = 3.14 * r * r;
     printf("Area: %.2f",area);
     return 0;
 }
